add healthchecks for container manager and segmentstore container registry
